new_poj/pyproject.toml
[tool.poetry]
name = "delivery"
version = "0.1.0"
description = ""
authors = ["Denis Kozlov <deniskozlow01.gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.8.2"
fastapi = "^0.112.0"
asyncio = "^3.4.3"
sqlalchemy = "^2.0.32"
pydantic-settings = "^2.4.0"
aiosqlite = "^0.20.0"
dynaconf = "^3.2.6"
uvicorn = "^0.30.6"
greenlet = "^3.0.3"
psycopg = {extras = ["binary"], version = "^3.2.1"}
asyncpg = "^0.29.0"
itsdangerous = "^2.2.0"
httpx = "^0.27.0"
fastapi-cache2 = "^0.2.2"
aioredis = "^2.0.1"
redis = "^5.0.8"
celery = "^5.4.0"
aiocache = "^0.12.2"
python-redis-cache = "^4.0.0"
jinja2 = "^3.1.4"
python-multipart = "^0.0.9"
pytest = "^8.3.2"
pytest-asyncio = "^0.24.0"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
